% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceldata_prep.R
\name{acceldata_prep}
\alias{acceldata_prep}
\title{Prepare 1-sec epoch accelerometer files  for processing}
\usage{
acceldata_prep(input_dir, file_pattern, output_dir, subjectID_length = 2,
  subjectID_start = 1)
}
\arguments{
\item{input_dir}{character sting of path to directory with original 1-sec
epoch accelerometer files}

\item{file_pattern}{charcter sting of naming convention for identifying all
original accelerometer files (equivalent to 'pattern' argument of
list.files() function)}

\item{output_dir}{charcter sting of path to directory where all the new
60-sec epoch 'PrepedData' files will be exported}

\item{subjectID_length}{numeric value of number of characters comprising the
unique subject identifier in the names of the original 1-sec accelerometer
files (e.g., if naming convention is "data_01" then set to 2, if naming
convention is "XXXX_data" then set to 4)}

\item{subjectID_start}{numeric value of the character in the names of the
original 1-sec accelerometer files where the unique subject identifier
starts (e.g., if naming convention is "data_01" then set to 6, if naming
convention is "XXXX_data" then set to 1)}
}
\value{
writes a CSV file ('PrepedData_ID.csv') in output_dir for each
original 1-sec epoch accelerometer file in the input_dir. Each new file is
in 60-sec epchs and contains the following new columns: TimeStamp (POSIXlt
object combining date and time), days (numeric value indicating which day of
monitoring each epoch is in, changing at each occurance of midnight),
wearing (character string of "w" when epoch is considered wear time and
"nw" for nonwear), and weekday (character string indicating day of the week)
}
\description{
Imports each 1-sec epoch accelerometer file (direct ActiLife export file) in
a directory and adds a time stamp column (combination of date and time),
reintegrates it to 60-sec epochs, add a 'day' indicator as a new column,
classifies wear time (using Choi 2011 algorithm) and exports new files to a
directory with the names 'PrepedData_ID' where 'ID' is the unique subject
identifier
}
\examples{
 acceldata_prep(input_dir = "./Data", file_pattern = "*1secAGdata.csv",
                output_dir = "./Outputs", subjectID_length = 2,
                subjectID_start = 1)
}
