% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_plot.R
\name{mvpa_plot}
\alias{mvpa_plot}
\title{Plot Weekday vs Weekend MVPA}
\usage{
mvpa_plot(dat, subject)
}
\arguments{
\item{subject}{id}

\item{a}{dataframe. x= weekday_classification, y = Average_MVPA}
}
\value{
a ggplot
}
\description{
Plot Weekday vs Weekend MVPA
}
\examples{
weekend_day_graph(subject1_data, S1)
}
