% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_plot_helper.R
\name{mvpa_plot_helper}
\alias{mvpa_plot_helper}
\title{Data arrangement for weekday vs weekend MVPA}
\usage{
mvpa_plot_helper(dat)
}
\arguments{
\item{data}{frame}
}
\value{
summary data frame of weekend/weeday MVPA including N, Average_MVPA, sd, se
}
\description{
Data arrangement for weekday vs weekend MVPA
}
\examples{
weekday_weekend_dat_arrange(my_data)
}
