% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pai_week_summary.R
\name{pai_week_summary}
\alias{pai_week_summary}
\title{Classify physical activity intensity from data frames in output folder}
\usage{
pai_week_summary(acceldata_PAI)
}
\arguments{
\item{acceldata_PAI}{use markPAI function from 'PhysicalActivity' package to assign character string
of physical activity intensities within "data"}

\item{data}{pull in data frame from output folder}

\item{intensity_week_summary}{master function used to create functions "pai_byday", "pai_byweekday_end",
and
"pai_weekavg"}

\item{pai_byday}{function used to create data frame classifying average physical activity intensity levels
within "sedentary", "light", "moderate", and "vigorous" categories for each day of the week}

\item{pai_byweekday_end}{function used to create data frame classifying physical activity intensity
as "sedentary, light, moderate, or vigorous" for the average of all weekdays compared to the
average of all weekend days}

\item{pai_weekavg}{function used to create data frame classifying physical activity intensity
as "sedentary, light, moderate, or vigorous" for the average of the entire week as a whole}
}
\value{
function "intensity_week_summary" computes values and associated data frames for variables
 "pai_byday", "pai_byweekday_end", and "pai_weekavg"
}
\description{
Classify physical activity intensity from data frames in output folder
}
\examples{
data <- read.csv("./Outputs/FilteredWakeData_CM.csv")

 acceldata_PAI <- markPAI(data = data, cts = "axis1", markingString = "w",
 breaks = c(-Inf, 100, 760, 2020, Inf),
 labels = c("sedentary", "light", "moderate", "vigorous"))

 intensity_week_summary <- function(acceldata_PAI)
}
