% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weekday_vs_end.R
\docType{data}
\name{pai_weekdayend}
\alias{pai_weekdayend}
\title{Plot Weekday vs Weekend MVPA}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8 rows and 3 columns.}
\usage{
pai_weekdayend
}
\arguments{
\item{pai_weekdayend}{dataframe or tible with summary of average minutes/day
 at physical activity intensities for weekdays vs the weekend

 @param steps_weekdayend dataframe or tible with summary of average
 steps/day physical activity intensities for weekdays vs the weekend

 @param subject unique subject identifier}
}
\value{
Two bar plot, one with weekday vs weekend average minutes/day at
physical activity intensites and the other with weekday vs weekend average
minutes/day
}
\description{
Plot Weekday vs Weekend MVPA
}
\examples{
plot_weekDay_vs_End(pai_summary, steps_summary, "S1")
}
\keyword{datasets}
