% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_steps.R
\name{plot_steps}
\alias{plot_steps}
\title{Plot steps per minute}
\usage{
plot_steps(pa.dat)
}
\arguments{
\item{a}{dataframe. x = study_min, y = steps}
}
\description{
Plot steps per minute
}
\examples{
plot_steps(my_data)
}
