% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weekday_vs_end.R
\name{plot_weekday_vs_end}
\alias{plot_weekday_vs_end}
\title{Plot average physical activity minutes/day and steps/day for weekday vs weekend}
\usage{
plot_weekday_vs_end(pai_weekdayend, steps_weekdayend, subject)
}
\arguments{
\item{pai_weekdayend}{dataframe or tible with summary of average minutes/day
at physical activity intensities for weekdays vs the weekend}

\item{steps_weekdayend}{dataframe or tible with summary of average
steps/day physical activity intensities for weekdays vs the weekend}

\item{subject}{unique subject identifier}
}
\value{
Two bar plot, one with weekday vs weekend average minutes/day at
physical activity intensites and the other with weekday vs weekend average
minutes/day
}
\description{
Plot average physical activity minutes/day and steps/day for weekday vs weekend
}
\examples{
plot_weekday_vs_end(pai_summary, steps_summary, "S1")
}
