% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_filter.R
\name{sleep_filter}
\alias{sleep_filter}
\title{Use a sleep log to remove accelerometer data collected during sleep}
\usage{
sleep_filter(acceldata_dir, sleeplog_dir, output_dir, subjectID_length,
  accelfile_pattern = "^PrepedData", sleeplog_pattern = "*SleepLog",
  accelfile_IDstart = 12, sleeplog_IDstart = 1)
}
\arguments{
\item{acceldata_dir}{character sting of path to directory with accelerometer
files}

\item{sleeplog_dir}{character sting of path to directory with sleep log files}

\item{output_dir}{charcter sting of path to directory where all the new
'FilteredWakeData' files will be exported}

\item{subjectID_length}{numeric value of number of characters comprising the
unique subject identifier in the names of the accelerometer and sleep log
files (e.g. if naming convention is "PrepedData_01" then set to 2, if naming
convention is "XXXX_data" then set to 4)}

\item{accelfile_pattern}{charcter sting of naming convention for identifying
all accelerometer files in acceldata_dir (equivalent to 'pattern' argument
of list.files() function)}

\item{sleeplog_pattern}{charcter sting of naming convention for identifying
all sleep log files in sleeplog_dir (equivalent to 'pattern' argument
of list.files() function)}

\item{accelfile_IDstart}{numeric value of the character in the names of the
accelerometer files in acceldata_dir where the unique subject identifier
 starts (e.g., if naming convention is "PrepedData_01" then set to 12, if
 naming convention is"XXXX_data" then set to 1)}

\item{sleeplog_IDstart}{numeric value of the character in the names of the
sleep log files in sleeplog_dir where the unique subject identifier starts
(e.g., if naming convention is "01_sleeplog_01" then set to 1, if naming
convention is "sleeplog_XXXX" then set to 1)}
}
\value{
Writes a CSV file ('FilteredWakeData_ID.csv') in output_dir for each
accelerometer file in the acceldata_dir. Each new file is the same as the
input accelerometer file except all epochs occuring between the bed and
wake time of each day (as indicated in the sleep log) are removed
}
\description{
Imports the sleep log and accelerometer physical activity file for each
subject, removes all epochs occuring between the bed and wake time of each
day (as indicated in the sleep log), and exports new files to a directory
with the names 'FilteredWakeData_ID' where 'ID' is the unique subject
identifier. Each accelerometer physical activity file must contain a
'TimeStamp' column with POSIXlt object of date and time for each epoch)
}
\examples{
 sleep_filter(acceldata_dir = "./Outputs", sleeplog_dir = "./Data",
             output_dir = "./Outputs", subjectID_length = 2,
             accelfile_pattern = "^PrepedData", sleeplog_pattern = "*SleepLog",
             accelfile_IDstart = 12, sleeplog_IDstart = 1)
}
