% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps_week_summary.R
\name{steps_week_summary}
\alias{steps_week_summary}
\title{Classify physical activity intensity from data frames in output folder}
\usage{
steps_week_summary(acceldata_PAI)
}
\arguments{
\item{acceldata_PAI}{1-minute epoch accelerometer data with sleep time
filtered out, a time stamp, and wear time classified (i.e., output from
sleep_filter function)}
}
\value{
returns three objects, each consisting of a dataframe: "steps_byday"
(steps/day for each day), "steps_byweekday_end" (average steps/day for
weekdays and for weekends), and "steps_weekavg" (average steps/day for the
entire week)
}
\description{
Classify physical activity intensity from data frames in output folder
}
\examples{
steps_week_summary(myPAdata)
}
