% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps_week_summary.R
\name{steps_week_summary}
\alias{steps_week_summary}
\title{Classify physical activity intensity from data frames in output folder}
\usage{
steps_week_summary(acceldata_PAI)
}
\arguments{
\item{acceldata_PAI}{function created within 'pai_week_changes' to assign physical activity intensity}

\item{data}{pull in data frame from output folder}

\item{steps_week_summary}{master function used to create functions 
"steps_byday", "steps_byweekday_end", and "steps_weekavg"}

\item{steps_byday}{function used to create data frame classifying average step counts
for each day of the week}

\item{pai_byweekday_end}{function used to create data frame classifying step counts
for the average of all weekdays compared to the average of all weekend days}

\item{pai_weekavg}{function used to create data frame classifying step counts
for the average of the entire week as a whole}
}
\value{
function "steps_week_summary" computes values and associated data frames for variables 
 "steps_byday", "steps_byweekday_end", and "steps_weekavg"
}
\description{
Classify physical activity intensity from data frames in output folder
}
\examples{
data <- read.csv("./Outputs/FilteredWakeData_CM.csv")

 acceldata_PAI <- markPAI(data = data, cts = "axis1", markingString = "w",
 breaks = c(-Inf, 100, 760, 2020, Inf),
 labels = c("sedentary", "light", "moderate", "vigorous"))
 
 steps_week_summary <- function(acceldata_PAI)
}
